pyproject.toml
[tool.poetry]
name = "fashionback"
version = "0.1.0"
description = ""
authors = ["Matvey <jefflol003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
sqlalchemy-utils = "^0.41.2"
asyncpg = "^0.29.0"
alembic = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
